[tests] Remove spurious error log in p2p_segwit.py
Since 88531390e670081b9b33e30dcd2083e365ca9fc4, when wait_until() fails,
an error message is logged to the test framework log. This means that if
wait_until() is called inside a try-except with the expectation that it
will fail, a spurious error message is logged.

wait_until() shouldn't be called with the expectation of failure. Fix
that in p2p_segwit.py.