Boost 1.71 and related upstream depends changes (#1093)
* depends: fix boost mac cross build with clang 9+

The ancient "darwin-4.9.1" profile has long been used to match against
clang, which prior to version 9, reported 4.9.1 as its version when
invoking "clang++ -dumpversion". Presumably this was a historical
compatibility quirk related to Apple's switch from gcc to clang.

This was "fixed" in clang 9.0, so that -dumpversion reports the real
version. Unfortunately that had the side-effect of breaking the
(brittle) boost compiler detection.

Move to the seemingly more-correct "clang-darwin" profile, which passes
the checks and builds correctly.

Also switch to using ar rather than libtool for archiving, as it's what
the clang-darwin profile expects to be using.

Note that because this is using a different profile, some of the final
command-line arguments end up changing. The changes look sane at a
glance.

* depends: Consistent use of package variable

All other mk files use the package variable consistently except for the two instances here, which have always been here, since depends was introduced in 0.10.

* depends: Propagate well-known vars into depends

For example, doing:

  make CC=clang CXX=clang++

Should now propagate these settings down to depends packages

* depends: boost: Specify toolset to bootstrap.sh

b2 will pickup our user-config.jam just fine, however, bootstrap.sh has
its own toolset autodetect mechanism, which doesn't GAF about our
user-config.jam

* depends: boost: Split target-os from toolset

Previously, we specified the target-os in the toolset (and sometimes
used the wrong command line flags), now we have a clear separation,
which is favored by ./bootstrap.sh and ./b2.

This means that all supported OSes will specify the correct target-os=
and toolset= on the command line.

* depends: boost: Use clang toolset if clang in CXX

* depends: Propagate only specific CLI variables to sub-makes

We want to supply well-known vars to ./configure scripts to do with as
they please. However, we do _not_ want to override these well-known vars
at make-time as certain build systems expect a self-mangled version of
these well-known vars.

For example, freetype and bdb will prepend `libtool --mode=compile' to
CC and CXX, which, if we override CC on the command line at make-time,
will break the build.

* depends: boost: Refer to version in URL

* depends: boost: Bump to 1.71.0

* depends: boost: Split into non-/native packages

* depends: boost: Disable all compression

* depends: boost: Cleanup architecture/address-model

* depends: boost: Cleanup toolset selection

* depends: boost: Remove unnecessary _archiver_

We already have $(package)_ar, so just use that instead

* depends: boost: Specify cflags+compileflags

* build,boost: update download url.

    - bintray is closing.
    - updated to jfrog.io.

* build: build Boost with -fcf-protection when targeting Darwin

The LLVM Clang we use for cross-compilation supports this option, and it's expected
that any builders on macOS will also be using an Apple Clang that supports it.

* build: don't use cf-protection when targeting arm-apple-darwin

After two reports on IRC of issues building depends on an Apple M1
machine, it turns out that this option can't be used when targeting
arm-apple-darwin. For now, just use it for x86_64-apple-darwin.

```bash
Apple clang version 12.0.5 (clang-1205.0.22.9)
Target: x86_64-apple-darwin20.4.0

error: option 'cf-protection=return' cannot be specified on this target
error: option 'cf-protection=branch' cannot be specified on this target
2 errors generated.
```

* build: fix unoptimized libraries in depends

We need to append-to rather than set CXXFLAGS, otherwise we loose -O2 &
-pipe. Currently this results in zeromq being built without optimizations
at all (or whatever the compiler would default too, essentially always -O0).

Bdb is the same, for the CXX portion of its code. C code has been built
with -O2.

Boost has actually been uneffected because it receives -O3 from it's own
build flags.

* build: don't install Boost cmake config files

* build: remove duplicate -fvisibility=hidden from Boost build

Boost already sets this by default.

Co-authored-by: Cory Fields <cory-nospam-@coryfields.com>
Co-authored-by: Carl Dong <contact@carldong.me>
Co-authored-by: fdov <fd21@pm.me>
Co-authored-by: fanquake <fanquake@gmail.com>